+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/changerequest/domain/service/ChangeRequestStatusService.java
@@ -67,7 +67,7 @@ public class ChangeRequestStatusService {
         ChangeRequestStatus crStatus = new ChangeRequestStatus();
         crStatus.setStatusChangeDateTime(LocalDateTime.now());
         crStatus.setChangeRequestStatusType(crStatusType);
-        crStatus.setUserPermission(resourcePermissionsFactory.get().getRoleByUser(getUserById((Long) AuthUtil.getCurrentUser().getId())));
+        crStatus.setUserPermission(resourcePermissionsFactory.get().getRoleByUser(getUserById(AuthUtil.getCurrentUser().getId())));
 
         return crStatusDAO.create(cr, crStatus);
     }
@@ -149,7 +149,7 @@ public class ChangeRequestStatusService {
         crStatus.setChangeRequestStatusType(crStatusType);
         crStatus.setComment(comment);
         crStatus.setStatusChangeDateTime(LocalDateTime.now());
-        crStatus.setUserPermission(resourcePermissionsFactory.get().getRoleByUser(getUserById((Long) AuthUtil.getCurrentUser().getId())));
+        crStatus.setUserPermission(resourcePermissionsFactory.get().getRoleByUser(getUserById(AuthUtil.getCurrentUser().getId())));
         if (resourcePermissionsFactory.get().isUserRoleAcbAdmin()) {
             crStatus.setCertificationBody(getCertificationBodyForCurrentUser());
         }
